refactor: extract SearchBar from ProjectListLayout
